requirements.txt
dotmap
eth-brownie
pytest
pandas
matplotlib==3.5.2
seaborn==0.12.1
rich==9.3.0
